refactor:add fixes to tasks of 'Exception handling'
