+++ src/main/java/com/rakovets/course/java/core/practice/exception_handling/tasks/Task01.java
@@ -1,4 +1,4 @@
-package com.rakovets.course.java.core.example.exception_handling.tasks;
+package com.rakovets.course.java.core.practice.exception_handling.tasks;
 
 public class Task01 {
     public static String getSubstring(String str) {
