+++ src/main/java/com/rakovets/course/java/core/practice/exception_handling/tasks/Task02.java
@@ -1,4 +1,4 @@
-package com.rakovets.course.java.core.example.exception_handling.tasks;
+package com.rakovets.course.java.core.practice.exception_handling.tasks;
 
 public class Task02 {
     public static String getIndexOf(int index, int[] array) {
