+++ src/test/java/com/rakovets/course/java/core/practice/exception_handling/demo/Task01Demo.java
@@ -1,6 +1,6 @@
 package com.rakovets.course.java.core.practice.exception_handling.demo;
 
-import com.rakovets.course.java.core.example.exception_handling.tasks.Task01;
+import com.rakovets.course.java.core.practice.exception_handling.tasks.Task01;
 
 public class Task01Demo {
     public static void main(String[] args) {
